refactor: deprecate CORSMiddleware from public interface.
It is used internally, its use is hardcoded and it doesn't work similar to other middleware when added to layers.

Should be implementation detail IMO.

Closes #3403
